app/services/data_enrichment.py
import aiohttp
from urllib.parse import urlparse
from bs4 import BeautifulSoup
from app.models.database import SessionLocal, BreachEntry

async def check_url_accessibility(url: str) -> dict:
    """Check if URL is accessible"""
    # TODO: Implement URL accessibility check
    return {
        "status_code": None,
        "title": None,
        "has_login_form": False
    }

async def resolve_domain(domain: str) -> dict:
    """Resolve domain to IP address"""
    # TODO: Implement domain resolution
    return {
        "resolved": False,
        "ip": None
    }

async def enrich_breach_data(entry_id: int):
    """Enrich breach data with additional information"""
    # TODO: Implement data enrichment
    pass 