requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
pydantic>=2.4.2
websockets>=12.0
requests>=2.31.0
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
groq>=0.3.0
python-multipart>=0.0.6
dnspython>=2.4.2
aiomysql>=0.2.0
cryptography>=41.0.7
tqdm>=4.66.1 